+++ config/config.php
@@ -1,0 +1,6 @@
+<?php
+define('TEMPLATES_DIR', '../templates/');
+define('LAYOUTS_DIR', 'layouts/');
+define('GALLERY_DIR', './img');
+require_once "../engine/functions.php";
+
