+++ templates/gallery.php
@@ -1,0 +1,2 @@
+<?php
+echo $gallery;
